# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Reliable Checkout and Rewards API
  version: 1.0.0
  description: |
    Cart, transactional checkout, coupon, and reporting API. Money fields are
    integer USD cents. `/admin/*` routes are administrative; authentication is
    intentionally out of scope for this assignment.
servers:
  - url: http://localhost:8080
tags:
  - name: Service
  - name: Products
  - name: Carts
  - name: Orders
  - name: Administration
paths:
  /health:
    get:
      tags: [Service]
      summary: Check process and database readiness
      responses:
        "200":
          description: Ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: ok }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /openapi.yaml:
    get:
      tags: [Service]
      summary: Return this OpenAPI contract
      responses:
        "200":
          description: The contract as YAML
          content:
            application/yaml:
              schema: { type: string }
  /docs:
    get:
      tags: [Service]
      summary: Interactive API reference rendered with Scalar
      responses:
        "200":
          description: HTML page that loads /openapi.yaml
          content:
            text/html:
              schema: { type: string }
  /products:
    get:
      tags: [Products]
      summary: List products
      responses:
        "200":
          description: Current product catalog and availability
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
  /carts:
    post:
      tags: [Carts]
      summary: Create an empty cart
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "201":
          description: Cart created
          headers:
            Location: { schema: { type: string }, description: Cart resource path }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /carts/{cartID}:
    parameters:
      - $ref: "#/components/parameters/CartID"
    get:
      tags: [Carts]
      summary: Get a cart using current catalog prices
      responses:
        "200":
          description: Cart found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /carts/{cartID}/items:
    parameters:
      - $ref: "#/components/parameters/CartID"
    post:
      tags: [Carts]
      summary: Add a product to a cart
      description: Returns a conflict if the product is already present; use PUT to change its quantity.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [product_id, quantity]
              properties:
                product_id: { type: integer, format: int64, minimum: 1 }
                quantity: { type: integer, minimum: 1, maximum: 1000000 }
      responses:
        "201":
          description: Item added; returns the updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "503": { $ref: "#/components/responses/RetryableConflict" }
  /carts/{cartID}/items/{productID}:
    parameters:
      - $ref: "#/components/parameters/CartID"
      - name: productID
        in: path
        required: true
        schema: { type: integer, format: int64, minimum: 1 }
    put:
      tags: [Carts]
      summary: Replace a cart item quantity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1, maximum: 1000000 }
      responses:
        "200":
          description: Quantity changed; returns the updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "503": { $ref: "#/components/responses/RetryableConflict" }
    delete:
      tags: [Carts]
      summary: Remove a cart item
      responses:
        "204": { description: Item removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "503": { $ref: "#/components/responses/RetryableConflict" }
  /carts/{cartID}/checkout:
    parameters:
      - $ref: "#/components/parameters/CartID"
      - name: Idempotency-Key
        in: header
        required: true
        description: Globally unique client-generated key, reused for an exact retry.
        schema: { type: string, minLength: 1, maxLength: 200 }
    post:
      tags: [Orders]
      summary: Atomically check out a cart
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                coupon_code:
                  type: string
                  description: Case-insensitive; surrounding whitespace is ignored.
      responses:
        "201":
          description: Order created or the completed order returned for an exact retry
          headers:
            Location: { schema: { type: string }, description: Order resource path }
            Idempotent-Replay:
              schema: { type: string, const: "true" }
              description: Present only when this response replays an existing order.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "422":
          description: The cart is empty
          content:
            application/json:
              schema: { $ref: "#/components/schemas/APIError" }
        "503": { $ref: "#/components/responses/RetryableConflict" }
  /orders/{orderID}:
    parameters:
      - name: orderID
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Orders]
      summary: Retrieve an immutable order snapshot
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/coupons:
    post:
      tags: [Administration]
      summary: Generate one coupon for the oldest unrewarded eligible milestone
      requestBody:
        required: false
        content:
          application/json:
            schema: { type: object, additionalProperties: false }
      responses:
        "201":
          description: Coupon generated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Coupon" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409":
          description: No unrewarded milestone is eligible
          content:
            application/json:
              schema: { $ref: "#/components/schemas/APIError" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "503": { $ref: "#/components/responses/RetryableConflict" }
  /admin/report:
    get:
      tags: [Administration]
      summary: Return a read-only point-in-time report
      responses:
        "200":
          description: Business report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }
components:
  parameters:
    CartID:
      name: cartID
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Malformed JSON, path parameter, or input value
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    Conflict:
      description: Operation conflicts with current cart, stock, idempotency, or coupon state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    PayloadTooLarge:
      description: Request body exceeds 1 MiB (`PAYLOAD_TOO_LARGE`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    RetryableConflict:
      description: |
        `RETRYABLE_CONFLICT`: the request waited too long for a lock, or hit a
        deadlock or serialization failure, and was rolled back without effect.
        Retry after the `Retry-After` delay; checkout retries must reuse the
        same `Idempotency-Key`.
      headers:
        Retry-After: { schema: { type: integer }, description: Seconds to wait before retrying }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    ServiceUnavailable:
      description: Database is unavailable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
  schemas:
    Product:
      type: object
      required: [id, name, unit_price_cents, inventory, currency]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        unit_price_cents: { type: integer, format: int64, minimum: 0 }
        inventory: { type: integer, minimum: 0 }
        currency: { type: string, const: USD }
    CartItem:
      type: object
      required: [product_id, product_name, quantity, unit_price_cents, line_total_cents]
      properties:
        product_id: { type: integer, format: int64 }
        product_name: { type: string }
        quantity: { type: integer, minimum: 1 }
        unit_price_cents: { type: integer, format: int64, minimum: 0 }
        line_total_cents: { type: integer, format: int64, minimum: 0 }
    Cart:
      type: object
      required: [id, status, items, subtotal_cents, currency, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [open, checked_out] }
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        subtotal_cents: { type: integer, format: int64, minimum: 0 }
        currency: { type: string, const: USD }
        order_id:
          type: string
          format: uuid
          description: Present once the cart is checked out; the order holds the prices actually charged.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrderItem:
      allOf:
        - $ref: "#/components/schemas/CartItem"
    Order:
      type: object
      required: [id, cart_id, items, subtotal_cents, discount_cents, total_cents, currency, idempotency_key, created_at]
      properties:
        id: { type: string, format: uuid }
        cart_id: { type: string, format: uuid }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        subtotal_cents: { type: integer, format: int64, minimum: 0 }
        discount_cents: { type: integer, format: int64, minimum: 0 }
        total_cents: { type: integer, format: int64, minimum: 0 }
        currency: { type: string, const: USD }
        coupon_code: { type: string }
        idempotency_key: { type: string }
        created_at: { type: string, format: date-time }
    Coupon:
      type: object
      required: [id, code, milestone_order_count, discount_percent, status, created_at]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        milestone_order_count: { type: integer, format: int64, minimum: 1 }
        discount_percent: { type: integer, minimum: 1, maximum: 100 }
        status: { type: string, enum: [available, redeemed] }
        redeemed_order_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        redeemed_at: { type: string, format: date-time }
    Report:
      type: object
      required: [purchased_quantity_by_product, gross_revenue_cents, total_discounts_cents, net_revenue_cents, currency, coupons, total_orders]
      properties:
        purchased_quantity_by_product:
          type: array
          items:
            type: object
            required: [product_id, product_name, quantity]
            properties:
              product_id: { type: integer, format: int64 }
              product_name: { type: string }
              quantity: { type: integer, format: int64, minimum: 1 }
        gross_revenue_cents: { type: integer, format: int64, minimum: 0 }
        total_discounts_cents: { type: integer, format: int64, minimum: 0 }
        net_revenue_cents: { type: integer, format: int64, minimum: 0 }
        currency: { type: string, const: USD }
        coupons:
          type: object
          required: [generated, available, redeemed, items]
          properties:
            generated: { type: integer, format: int64, minimum: 0 }
            available: { type: integer, format: int64, minimum: 0 }
            redeemed: { type: integer, format: int64, minimum: 0 }
            items:
              type: array
              description: Every generated coupon with its status, ordered by milestone.
              items: { $ref: "#/components/schemas/Coupon" }
        total_orders: { type: integer, format: int64, minimum: 0 }
    APIError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: INSUFFICIENT_INVENTORY }
            message: { type: string }